academics/requirements.txt
Django>=4.2
djangorestframework
gunicorn
whitenoise
django-cors-headers
psycopg2-binary
dj-database-url
python-dotenv
